ivy: update to math/rand/v2
Because math/big is not updated to v2, we must do a bit more of the lifting
ourselves, but the word-stuffing interface to big.Int makes it pretty easy.

One minor annoyance, probably not worth fixing, is that the Seed for PCG
is two words, but the UI we have only allows one.

Fixes #174

+++ parse/special.go
@@ -329,7 +329,7 @@ Switch:
 			p.Println(conf.RandomSeed())
 			break Switch
 		}
-		conf.SetRandomSeed(p.nextDecimalNumber64())
+		conf.SetRandomSeed(uint64(p.nextDecimalNumber64()))
 	case "timezone":
 		if p.peek().Type == scan.EOF {
 			_, offset := time.Now().In(conf.Location()).Zone()
